feat: allow configurable number of days for astro table
The number of days displayed in the sunrise/sunset (astro) table was
previously tied to the number of days available in the 3-hour forecast
data. This meant the astro table could not show more days than the
forecast snapshot.

This change makes the number of days in the astro table configurable via
the daily number of days setting.

- The tooltip header has been updated to "Next 3 Day(s) Snapshot" to
reflect the potential for a variable number of days.